cleaner way to determine sampling method
